backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
asyncpg==0.29.0
sqlmodel==0.0.14
sqlalchemy[asyncio]==2.0.25
pgvector==0.2.4

# PDF processing
PyPDF2==3.0.1

# HTTP client
httpx==0.26.0

# Settings management
pydantic-settings==2.1.0

# Multipart form data
python-multipart==0.0.9

# Observability
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Resilience
tenacity==8.2.3

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx>=0.26.0
